bugfix
Using object pooling together with the new job system introduced a gotcha.
When the same object in the pool is reused, we must not add the object's transform into the job queue again. This will cause the movement job to be run on the object's transform multiple times causing it to go faster every time it is instantiated from the pool.